Mock data tweaks
