+++ mock-api/snapshot.ts
@@ -14,7 +14,7 @@ import { GiB } from '~/util/units'
 
 import { disks } from './disk'
 import type { Json } from './json-type'
-import { Rando } from './msw/util'
+import { Rando } from './msw/rando'
 import { project } from './project'
 
 // Use seeded random for consistent states across runs
